+++ models/utilities/bin_counter/src/bin_counter.cc
@@ -165,7 +165,7 @@ CML_BinCounter::set_data( double limit_a,
     bins[0].bin_floor = lower_limit;
     for (size_t ii = 0; ii < nbin-1; ii++) {
       bins[ii].bin_ceil    =
-      bins[ii+1].bin_floor = lower_limit + (ii+1) * interval;
+      bins[ii+1].bin_floor = lower_limit + static_cast<double>(ii+1) * interval;
       bins[ii].count = 0;
     }
     bins[nbin-1].bin_ceil = upper_limit;
@@ -178,7 +178,7 @@ CML_BinCounter::set_data( double limit_a,
     bins[0].bin_floor = std::numeric_limits<double>::lowest();
     for (size_t ii = 0; ii < nbin-1; ii++) {
       bins[ii].bin_ceil =
-      bins[ii+1].bin_floor = lower_limit + ii * interval;
+      bins[ii+1].bin_floor = lower_limit + static_cast<double>(ii) * interval;
       bins[ii].count = 0;
     }
     bins[nbin-1].bin_ceil = std::numeric_limits<double>::max();
@@ -206,8 +206,8 @@ CML_BinCounter::insert(double value)
   // Consider only values below upper edge and only if model passed sanity
   // check.
   if (bins_ready && value <= bins[nbin-1].bin_ceil) {
-    for (int ii = nbin - 1; ii >= 0; ii--) {
-      const size_t bin_index = static_cast<size_t>(ii);
+    for (size_t ii = nbin - 1; ii >= 0; ii--) {
+      const size_t bin_index = ii;
       if (value >= bins[bin_index].bin_floor) {
         bins[bin_index].count++;
         return;
